Refactor: URL 수정

+++ src/main/java/ticketaka/mtvs3_final_backend/ticketing/ticket/custom/query/dto/TicketCustomQueryResponseDTO.java
@@ -15,7 +15,7 @@ public class TicketCustomQueryResponseDTO {
     public record getTicketDTO(
             ticketConcertDTO concertInfo,
             String seatInfo,
-            Long ticketId,
+            int ticketId,
             byte[] ticketImage
     ) {
     }
