+++ src/main/java/ticketaka/mtvs3_final_backend/ticketing/ticket/custom/query/service/TicketCustomQueryService.java
@@ -86,7 +86,7 @@ public class TicketCustomQueryService {
                                             concert.getConcertDate().toLocalTime().toString()
                                     ),
                                     seatInfo,
-                                    ticket.getId(),
+                                    ticket.getId().intValue(),
                                     ticketImage
                             );
                         })
